BUG land_mask now works again for regional
I also cleaned up various parts of Region, including changing the
`land_mask` kwarg of __init__ to `do_land_mask`.  This is not backwards
compatible: you'll have to grep+replace in your region objects.  But I
think it's worth it, to make more clear that this is a flag about how to
handle the land mask, not the actual land mask itself.

I inadvertently included adding the {LAT,LON}_BOUNDS_STR constants in
this commit as well.  In a perfect world I would have done those separately.