Fix workaround for years less than 1678 for non-0001 reference dates (#189)
Fixes bugs in numpy_datetime_range_workaround

Switch to using the minimum and maximum years in the loaded dataset
to determine whether dates should be offset.  This is what ultimately
determines whether dates are offset in initially in
numpy_datetime_workaround_encode_cf, so it is important to be
consistent.

Originally this method used the user-input subset bounds.  This
would have the potential of being an issue if the dataset included
data from times before 1678, say 1600, but the user only wanted to subset
data between say 1679 and 1750.  In that case internally, dates
would be offset to start in 1678, but the subset dates would not be
offset.  In addition this method would not catch if dates were
outside the Timestamp upper bound.